# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Speed dating chatroom API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new speed dating room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/match:
    post:
      operationId: matchRoom
      tags: [rooms]
      summary: Auto-match a chooser with the top 5 compatible suitors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchInput"
      responses:
        "201":
          description: Room created and suitors matched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResult"
        "409":
          description: Not enough suitors in pool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotEnoughSuitors"
...
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rooms/{id}/join:
    post:
      operationId: joinRoom
      tags: [rooms]
      summary: Join a room as chooser or suitor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinInput"
      responses:
        "200":
          description: Joined successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinResult"
        "400":
          description: Room full or invalid role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rooms/{id}/messages:
    get:
      operationId: getRoomMessages
      tags: [rooms]
      summary: Get message history for the chooser's view (all suitor chats)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: All messages grouped by suitor
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
  /rooms/{id}/choose:
    post:
      operationId: chooseWinner
      tags: [rooms]
      summary: Chooser picks their favourite suitor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChooseInput"
      responses:
        "200":
          description: Winner chosen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rooms/{id}/questions/{questionId}/rating:
    post:
      operationId: rateRoomQuestion
      tags: [rooms]
      summary: Rate a question that appeared in the current room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: questionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateRoomQuestionInput"
      responses:
        "201":
          description: Question rating recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionRatingResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Room or question not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /rooms/{id}/eliminate:
    post:
      operationId: eliminateSuitor
      tags: [rooms]
      summary: Eliminate a suitor from the current round
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EliminateSuitorInput"
      responses:
        "200":
          description: Updated room after elimination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /rooms/{id}/advance-round:
    post:
      operationId: advanceRoomRound
      tags: [rooms]
      summary: Advance the room to the next round
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated room after advancing round
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    MatchInput:
      type: object
      properties:
        chooserUserId:
          type: string
        maxSuitors:
          type: integer
          minimum: 3
          maximum: 6
          description: "Optional number of suitors for this auto-match session."
        numberOfRounds:
          type: integer
          description: "Optional: choose 3 or 5 rounds for the session. Defaults to 3."
        questionConfig:
          $ref: "#/components/schemas/RoomQuestionConfig"
      required:
        - chooserUserId
    MatchResult:
      allOf:
        - $ref: "#/components/schemas/Room"
        - type: object
          properties:
            chooserParticipantId:
              type: string
          required:
            - chooserParticipantId
    NotEnoughSuitors:
      type: object
      properties:
        error:
          type: string
        count:
          type: integer
      required:
        - error
        - count
    Participant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [chooser, suitor]
        suitorSlot:
          type: ["integer", "null"]
        isBot:
          type: boolean
          description: "True when this participant slot is an AI bot contestant"
        isPremium:
          type: boolean
          description: "True when this participant holds an active premium subscription (server-validated)"
      required:
        - id
        - name
        - role
        - suitorSlot
        - isBot
        - isPremium
    Room:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        status:
          type: string
          enum: [waiting, active, ended]
        roomSnapshotVersion:
          type: integer
          description: "Monotonically increasing revision for canonical room snapshots."
        chooserName:
          type: ["string", "null"]
        suitorCount:
          type: integer
        maxSuitors:
          type: integer
        numberOfRounds:
          type: integer
          description: "Number of rounds for this session (defaults to 3 if absent)."
        roundDurationSeconds:
          type: integer
          description: "Per-round duration in seconds for chooser/question phase."
        answerTimeSeconds:
          type: integer
          description: "Answer time in seconds for suitor responses."
        roundEndsAt:
          type: ["string", "null"]
          description: "ISO timestamp when the current round ends (server authoritative)."
        currentRound:
          type: integer
          description: "Current round (1-4). Rounds 1-3: 1 question/suitor then eliminate. Round 4: 3 questions/suitor then choose winner."
        eliminatedParticipants:
          type: array
          items:
            type: string
          description: "Array of participant IDs who have been eliminated"
        winnerId:
          type: ["string", "null"]
        winnerName:
          type: ["string", "null"]
        participants:
          type: array
          items:
            $ref: "#/components/schemas/Participant"
        questionConfig:
          $ref: "#/components/schemas/RoomQuestionConfig"
        currentRoundQuestions:
          type: array
          items:
            $ref: "#/components/schemas/RoomRoundQuestion"
        createdAt:
          type: string
      required:
        - id
        - code
        - status
        - roomSnapshotVersion
        - chooserName
        - suitorCount
        - maxSuitors
        - currentRound
        - eliminatedParticipants
        - winnerId
        - winnerName
        - participants
        - questionConfig
        - currentRoundQuestions
        - createdAt
    Message:
      type: object
      properties:
        id:
          type: string
        roomId:
          type: string
        senderId:
          type: string
        senderName:
          type: string
        senderRole:
          type: string
          enum: [chooser, suitor]
        suitorSlot:
          type: ["integer", "null"]
        round:
          type: ["integer", "null"]
          description: "Which speed-dating round this message belongs to"
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - roomId
        - senderId
        - senderName
        - senderRole
        - suitorSlot
        - round
        - content
        - createdAt
    RoomInput:
      type: object
      properties:
        chooserName:
          type: string
          minLength: 1
        maxSuitors:
          type: integer
          minimum: 3
          maximum: 6
          description: "Optional number of suitors for this session. The game length will be maxSuitors - 1 rounds."
        numberOfRounds:
          type: integer
          description: "Optional: choose 3 or 5 rounds for the session. Defaults to 3."
        questionConfig:
          $ref: "#/components/schemas/RoomQuestionConfig"
      required:
        - chooserName
    QuestionCategory:
      type: string
      enum: [general, fun, deep]
    QuestionDifficulty:
      type: string
      enum: [easy, medium, hard]
    RoundDifficulty:
      type: string
      enum: [easy, medium, hard, wildcard]
    QuestionPackStrategy:
      oneOf:
        - type: object
          properties:
            mode:
              type: string
              enum: [single]
            packSlug:
              type: string
              minLength: 1
          required: [mode, packSlug]
        - type: object
          properties:
            mode:
              type: string
              enum: [multi]
            packSlugs:
              type: array
              minItems: 1
              items:
                type: string
                minLength: 1
          required: [mode, packSlugs]
        - type: object
          properties:
            mode:
              type: string
              enum: [weighted]
            packWeights:
              type: object
              additionalProperties:
                type: number
                minimum: 0
          required: [mode, packWeights]
        - type: object
          properties:
            mode:
              type: string
              enum: [seasonal]
            seasonalPackSlugs:
              type: array
              minItems: 1
              items:
                type: string
                minLength: 1
            fallbackPackSlugs:
              type: array
              items:
                type: string
                minLength: 1
          required: [mode, seasonalPackSlugs]
        - type: object
          properties:
            mode:
              type: string
              enum: [premium]
            premiumPackSlugs:
              type: array
              minItems: 1
              items:
                type: string
                minLength: 1
            fallbackPackSlugs:
              type: array
              minItems: 1
              items:
                type: string
                minLength: 1
          required: [mode, premiumPackSlugs, fallbackPackSlugs]
    RoomQuestionConfig:
      type: object
      properties:
        packStrategy:
          $ref: "#/components/schemas/QuestionPackStrategy"
        categoryWeights:
          type: object
          properties:
            general:
              type: number
              minimum: 0
            fun:
              type: number
              minimum: 0
            deep:
              type: number
              minimum: 0
        difficultyPlan:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/RoundDifficulty"
        avoidRecentGames:
          type: integer
          minimum: 0
          maximum: 20
        disallowRepeatsInGame:
          type: boolean
    RoomRoundQuestion:
      type: object
      required: [id, content, packSlug, category, difficulty, ratingSummary]
      properties:
        id:
          type: string
        content:
          type: string
        packSlug:
          type: string
        category:
          $ref: "#/components/schemas/QuestionCategory"
        difficulty:
          $ref: "#/components/schemas/QuestionDifficulty"
        ratingSummary:
          $ref: "#/components/schemas/QuestionRatingSummary"
    QuestionRatingSummary:
      type: object
      required: [questionId, ratingCount, averageRating, qualityScore]
      properties:
        questionId:
          type: string
        ratingCount:
          type: integer
        averageRating:
          type: ["number", "null"]
        qualityScore:
          type: ["number", "null"]
...
    RateRoomQuestionInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        participantId:
          type: string
          description: Optional participant id used to tie ratings to a live room seat.
    QuestionRatingResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    JoinInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
          enum: [chooser, suitor]
      required:
        - name
        - role
    JoinResult:
      type: object
      properties:
        participantId:
          type: string
        room:
          $ref: "#/components/schemas/Room"
      required:
        - participantId
        - room
    ChooseInput:
      type: object
      properties:
        winnerId:
          type: string
      required:
        - winnerId
...
    EliminateSuitorInput:
      type: object
      required: [participantId]
      properties:
        participantId:
          type: string
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
          enum: [unauthorized, forbidden, not_found, invalid_query, internal_error]
        details:
          type: object
          additionalProperties: true